AssistLink-Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic[email]>=2.10.0
pymongo>=4.5.0
motor>=3.3.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pyjwt>=2.8.0
